Add support for protected payloads
Add support for protected payloads. The new `intranpayload' option can
be used to specify a translation function translating payloads to
values of the translated type.  This function will be used instead of
the `intran' function to to look up the receiving object of a message
in a server.

This makes it easy to use the protected payloads introduced in GNU
Mach 1.5.

An inTransPayload function translates payloads to objects, like an
inTrans function translates from port names to objects.  Generate code
in the server routine to optimize lookups to the receiver of the
message.

Additionally, if no intran function is provided, but an intranpayload
function is, it is expected to translate from payloads to port names.
This is used to preserve the semantics in case the server routine
expects a port name.

* NEWS: Add item.
* lexxer.l: Emit syInTranPayload.
* parser.h: Define syInTranPayload.
* parser.y (TransTypeSpec): Handle syInTranPayload.
* type.h (struct ipc_type): Add itInTransPayload.
* server.c (WriteExtractArgValue): If a payload-aware intrans function
has been specified, use it to get a reference to the receiving object.
* routine.c (rtAugmentArgKind): Force the use of a local variable if a
payload-aware translate-in function is defined.