Check that msgt_name is always smaller than 255.
For the x86_64 ABI we want this to always fit into 1 byte. Even for
regular i686, msgt_name is always smaller than 25 (MACH_MSG_TYPE_LAST)
and we don't have plans to have more names.

Also throw an error if we deemed an RPC to be "TooLong" as that won't
work or work badly.

Tested by cross-compiling a basic Hurd system.
Message-Id: <ZFsuKtiLdwNpD6b1@jupiter.tail36e24.ts.net>